+++ app/routes/components/Field/changelog.codegen.ts
@@ -9,4 +9,5 @@ export const changelog = ''
   + '- 2025-11-21: Add `Field.FileInput` component. And remove useless style\n'
   + '  `width: 100%`.\n'
   + '- 2025-12-17: Export the `useOptionalFieldContext` hook.\n'
+  + '- 2026-02-14: Use a semantic `fieldset` wrapper and remove `biome-ignore`.\n'
   ;