+++ app/routes/components/ScrollArea/changelog.codegen.ts
@@ -13,4 +13,5 @@ export const changelog = ''
   + '- 2025-11-04: Support the `ScrollArea.ScrollbarWithThumb`.\n'
   + '- 2025-12-11: Support the horizontal scrollbars with prop\n'
   + '  `orientation="horizontal"`.\n'
+  + '- 2026-02-14: Remove unnecessary `biome-ignore` around root hover handlers.\n'
   ;