+++ src/components/DatePicker/index.ts
@@ -1,3 +1,3 @@
-// DatePicker from copy-ui @ 2026-02-13.
+// DatePicker from copy-ui @ 2026-02-14.
 
 export { DatePicker } from "./date-picker";