+++ src/components/ScrollArea/index.ts
@@ -1,3 +1,3 @@
-// ScrollArea from copy-ui @ 2025-12-11.
+// ScrollArea from copy-ui @ 2026-02-14.
 
 export { ScrollArea } from "./scroll-area";